fix(shaders): split iTime into hi/lo to prevent animation freeze at long uptime
float32 iTime grows with daemon uptime — around 36 h the ULP exceeds
one frame at 60 fps and any animation driven purely by iTime freezes.

Pipeline changes:
- Accumulate elapsed seconds as double through the entire C++ path
  (overlayservice → QML property → ZoneShaderItem → render node).
- At the final GPU upload, split into iTime (fmod residual in [0,1024))
  and iTimeHi (integer wrap offset). iTime stays precise; iTimeHi
  changes once per ~17 min and gets its own partial UBO upload.

GLSL helpers:
- timeSin(speed) / timeCos(speed) reconstruct continuous phase across
  the 1024 s wrap via angle-addition with mod-TAU reduction on the
  large-phase term. Offset overloads for per-instance hash phases.

Shader conversions (76 call sites, 11 files):
- All drift-family computeInstanceUV Lissajous drift, rotation, and
  breathe calls (cachyos, opensuse, tumbleweed, neon, fedora, nixos,
  arch, endeavouros). Berry-drift orbit and voxel-terrain camera roll.

+++ src/daemon/overlayservice/shader.cpp
@@ -228,12 +228,16 @@ void OverlayService::updateShaderUniforms()
         currentTime = m_shaderTimer.elapsed();
     }
 
-    const float iTime = currentTime / 1000.0f;
+    // Keep iTime as double through the whole pipeline. ZoneShaderNodeRhi splits
+    // it into iTime (wrapped) + iTimeHi at the final GPU upload — see
+    // kShaderTimeWrap. Casting to float32 here would requantize the counter
+    // before the wrap, reintroducing the freezing bug at long uptimes.
+    const double iTime = static_cast<double>(currentTime) / 1000.0;
 
     // Calculate delta time with clamp (max 100ms prevents jumps after sleep/resume)
     constexpr float maxDelta = 0.1f;
     const qint64 lastTime = m_lastFrameTime.exchange(currentTime);
-    float iTimeDelta = qMin((currentTime - lastTime) / 1000.0f, maxDelta);
+    float iTimeDelta = qMin(static_cast<float>(currentTime - lastTime) / 1000.0f, maxDelta);
 
     // Prevent frame counter overflow (reset at 1 billion, ~193 days at 60fps)
     int frame = m_frameCount.fetch_add(1);
